battleship-game/tests/test_game.py
import unittest
from src.game import Game
from src.player import Player
from src.gameboard import GameBoard

class TestGame(unittest.TestCase):
    def setUp(self):
        # Setup for testing; initialize players, boards, and a game instance
        size = 5
        self.gameboard_p1 = GameBoard(size)
        self.gameboard_p2 = GameBoard(size)

        # For simplicity, placing one ship for each player
        self.gameboard_p1.place_ship((1, 1))
        self.gameboard_p2.place_ship((2, 2))

        self.player1 = Player("Player 1", self.gameboard_p1)
        self.player2 = Player("Player 2", self.gameboard_p2)

        self.game = Game(size, [(1, 1)], [(2, 2)], [(2, 2)], [(1, 1)])  # Simplified setup

    def test_initialization(self):
        # Test game initialization
        self.assertEqual(self.game.board_size, 5)
        self.assertIsInstance(self.game.player1, Player)
        self.assertIsInstance(self.game.player2, Player)

    def test_determine_winner(self):
        # Simulate a scenario and test winner determination
        # Assuming both players hit each other's ship once
        self.game.player1.gameboard.receive_attack((2, 2))  # P1 hits P2
        self.game.player2.gameboard.receive_attack((1, 1))  # P2 hits P1

        # Manually call the method to determine winner for this test scenario
        winner = self.game._determine_winner()

        # Since both players hit once, it's a draw; adjust based on actual implementation
        self.assertEqual(winner, "It's a draw!")

    # Additional tests can be added here to cover more scenarios and methods

if __name__ == "__main__":
    unittest.main()
